Ensure `correct + wrong` is not 0 in reliable role
If there is an API issue, then the word will not be accepted by the bot. However, the database will already have a record for that user in which both `correct` and `wrong` are set to zero. This will throw a zero division error when trying to calculate the accuracy in `add_remove_reliable_role`. Hence, this check is necessary.