Handle `NotFound` error in `/leaderboard` cmd
If the retrieved list of top 10 users had someone who had left the server, the bot would throw an error. Fixed that by handling the exception. In addition, if an exception is raised, the bot will try to retrieve more users until the quota of 10 is filled, or until no more users are available in the db.